Reorganization of test functions into modules
