+++ examples/ga_ackley_test_function.rs
@@ -6,7 +6,7 @@ fn main() {
 
   let best_individual = ga::Builder::with_rvc()
     .set_max_duration(std::time::Duration::from_secs(5))
-    .fitness_fn(|x| -test_functions::ackley(x))
+    .fitness_fn(|x| -test_functions::multi_minima::ackley(x))
     .dim(4)
     .build()
     .run()
